Create auth middleware and use it for protected routes
